# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "summary": "Register a new user account",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created. Confirm email before logging in." },
          "401": { "description": "Signup failed (e.g., email already exists)" },
          "429": { "description": "Rate limit exceeded (5 req/min)" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Authenticate and receive session tokens",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns accessToken, refreshToken, expiresAt, and user metadata" },
          "401": { "description": "Invalid email or password" },
          "429": { "description": "Rate limit exceeded (5 req/min)" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Exchange refresh token for new access token",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns new accessToken and refreshToken" },
          "401": { "description": "Invalid or expired refresh token" },
          "429": { "description": "Rate limit exceeded (5 req/min)" }
        }
      }
    },
    "/api/auth/switch-tenant": {
      "post": {
        "operationId": "AuthController_switchTenant",
        "summary": "Switch active tenant context. Client MUST call /auth/refresh afterward.",
        "tags": ["Auth"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SwitchTenantDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Context switched. Call POST /auth/refresh for updated JWT." },
          "403": { "description": "Not a member of the requested tenant" }
        }
      }
    },
...
    "/api/posts/{postId}/comments": {
      "post": {
        "operationId": "CommentsController_createComment",
        "summary": "Create a comment on a post",
        "tags": ["Comments"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCommentDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Comment created. Notification sent to post author." },
          "404": { "description": "Post not found in current tenant" }
        }
      },
      "get": {
        "operationId": "CommentsController_getComments",
        "summary": "List comments for a post (newest first)",
        "tags": ["Comments"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "postId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Paginated array of comments" },
          "404": { "description": "Post not found in current tenant" }
        }
      }
    },
...
    "/api/users/{id}/follow": {
      "post": {
        "operationId": "FollowsController_follow",
        "summary": "Follow a user (platform-wide, not tenant-scoped)",
        "tags": ["Follows"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "201": { "description": "Now following the user" },
          "409": { "description": "Already following this user" }
        }
      },
      "delete": {
        "operationId": "FollowsController_unfollow",
        "summary": "Unfollow a user",
        "tags": ["Follows"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Unfollowed successfully" }
        }
      }
    },
    "/api/users/{id}/followers": {
      "get": {
        "operationId": "FollowsController_getFollowers",
        "summary": "List users who follow the specified user",
        "tags": ["Follows"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Paginated follower list" }
        }
      }
    },
    "/api/users/{id}/following": {
      "get": {
        "operationId": "FollowsController_getFollowing",
        "summary": "List users the specified user is following",
        "tags": ["Follows"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Paginated following list" }
        }
      }
    },
    "/api/channels": {
      "post": {
        "operationId": "ChatController_createChannel",
        "summary": "Create a chat channel (public, private, or direct)",
        "tags": ["Chat"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChannelDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Channel created. Creator auto-added as member." }
        }
      },
      "get": {
        "operationId": "ChatController_getChannels",
        "summary": "List accessible channels in the current tenant",
        "tags": ["Chat"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Array of channels" }
        }
      }
    },
    "/api/channels/{id}/members": {
      "post": {
        "operationId": "ChatController_addMember",
        "summary": "Add a member to a channel",
        "tags": ["Chat"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddMemberDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Member added (idempotent)" }
        }
      }
    },
    "/api/channels/{id}/messages": {
      "post": {
        "operationId": "ChatController_sendMessage",
        "summary": "Send a message to a channel",
        "tags": ["Chat"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMessageDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Message sent. Push notification dispatched for private/direct channels." }
        }
      },
      "get": {
        "operationId": "ChatController_getMessages",
        "summary": "Get channel messages with cursor-based pagination",
        "tags": ["Chat"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "UUID of last message seen — loads older messages" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "{ messages, nextCursor } — nextCursor is null when no more messages" }
        }
      }
    },
...
  },
  "info": {
    "title": "ChurchApp Platform API",
    "description": "Multi-tenant church management platform & global social network.\n\n**Authentication:** All protected endpoints require a Supabase JWT in the `Authorization: Bearer <token>` header.\n\n**Tenant Context:** Most endpoints are tenant-scoped via PostgreSQL Row-Level Security. Call `POST /api/auth/switch-tenant` followed by `POST /api/auth/refresh` to set the active tenant in your JWT.\n\n**Rate Limits:** Global: 100 req/min per IP. Auth endpoints: 5 req/min per IP. Webhooks: unlimited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase JWT access token"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 8, "maxLength": 72, "example": "secureP@ss1" }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "secureP@ss1" }
        }
      },
      "RefreshDto": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "description": "Refresh token from login or previous refresh" }
        }
      },
      "SwitchTenantDto": {
        "type": "object",
        "required": ["tenantId"],
        "properties": {
          "tenantId": { "type": "string", "format": "uuid", "description": "Target tenant to switch context to" }
        }
      },
...
      "CreateCommentDto": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string", "maxLength": 2000, "example": "Great post!" }
        }
      },
...
      "CreateChannelDto": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "name": { "type": "string", "maxLength": 100, "example": "Prayer Group" },
          "type": { "type": "string", "enum": ["public", "private", "direct"], "example": "private" }
        }
      },
      "AddMemberDto": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "type": "string", "format": "uuid", "description": "User ID to add to the channel" }
        }
      },
      "SendMessageDto": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string", "minLength": 1, "maxLength": 5000, "example": "Hello everyone!" }
        }
      },
...
    }
  }
}